Centric Patch - Gemeentepassen 1.4

+++ backend/local4local/src/main/java/nl/centric/innovation/local4local/dto/OfferTransactionInvoiceDto.java
@@ -1,11 +1,11 @@
 package nl.centric.innovation.local4local.dto;
 
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Builder;
 import lombok.NonNull;
 import nl.centric.innovation.local4local.entity.OfferTransaction;
 import nl.centric.innovation.local4local.util.DateUtils;
 
-import javax.validation.constraints.NotEmpty;
 
 @Builder
 public record OfferTransactionInvoiceDto(
@@ -20,7 +20,7 @@ public record OfferTransactionInvoiceDto(
 ) {
     public OfferTransactionInvoiceDto(String passNumber, OfferTransaction offerTransaction) {
         this(passNumber,
-                offerTransaction.getAmount(),
+                offerTransaction.getAmount().doubleValue(),
                 offerTransaction.getOfferBenefitName(),
                 DateUtils.formatDateDefault(offerTransaction.getCreatedDate().toLocalDate()));
     }
